Move factories in dedicated folder for clarity, split BaseGalaxyCatalog with SnapshotCatalog
